pyproject.toml
[tool.poetry]
name = "electronics-network"
version = "0.1.0"
description = ""
authors = ["macbook <kriukov@pos-credir.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.6"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-filter = "^25.1"
black = "^25.1.0"
flake8 = "^7.1.2"
mypy = "^1.15.0"
isort = "^6.0.0"
v = "^1"
djangorestframework-simplejwt = "^5.4.0"
drf-spectacular = "^0.28.0"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
